requirements.txt
aiofiles==0.6.0
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
fastapi==0.63.0
h11==0.12.0
html5lib==1.0.1
idna==2.8
imageio==2.9.0
ipaddr==2.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.5
packaging==20.3
pandas==1.2.0
pep517==0.8.2
Pillow==8.1.0
progress==1.5
pydantic==1.7.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
six==1.15.0
starlette==0.13.6
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.25.8
uvicorn==0.13.3
webencodings==0.5.1
